create produk & dark mode admin

+++ database/migrations/2020_07_16_062651_add_relationships_to_products_table.php
@@ -14,7 +14,7 @@ class AddRelationshipsToProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->integer('category_id')->unsigned()->change();
+            $table->BigInteger('category_id')->unsigned()->change();
             $table->foreign('category_id')->references('id')->on('categories')->onUpdte('cascade')->onDelete('cascade');
         });
     }
