+++ database/migrations/2020_07_16_063623_add_relationships_to_orders_table.php
@@ -14,10 +14,10 @@ class AddRelationshipsToOrdersTable extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->integer('customer_id')->unsigned()->change();
+            $table->BigInteger('customer_id')->unsigned()->change();
         $table->foreign('customer_id')->references('id')->on('customers')
             ->onUpdate('cascade')->onDelete('cascade');
-        $table->integer('user_id')->unsigned()->change();
+        $table->BigInteger('user_id')->unsigned()->change();
         $table->foreign('user_id')->references('id')->on('users')
             ->onUpdate('cascade')->onDelete('cascade');
         });
@@ -38,7 +38,7 @@ class AddRelationshipsToOrdersTable extends Migration
                 $table->dropIndex('orders_customer_id_foreign');
             });
             Schema::table('orders', function(Blueprint $table) {
-                $table->integer('customer_id')->change();
+                $table->BigInteger('customer_id')->change();
             });
             Schema::table('orders', function(Blueprint $table) {
                 $table->dropForeign('orders_user_id_foreign');
@@ -47,7 +47,7 @@ class AddRelationshipsToOrdersTable extends Migration
                 $table->dropIndex('orders_user_id_foreign');
             });
             Schema::table('orders', function(Blueprint $table) {
-                $table->integer('user_id')->change();
+                $table->BigInteger('user_id')->change();
             });
         });
     }
